+++ aiohttp_utils/__init__.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
-__version__ = '1.0.0'
-__author__ = 'Steven Loria'
-__license__ = "MIT"
-
 from .negotiation import Response
 from .constants import APP_KEY, CONFIG_KEY
 from .runner import run
+
+__version__ = '2.0.0.dev0'
+__author__ = 'Steven Loria'
+__license__ = "MIT"
 
 __all__ = (
     'Response',