+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 *********
 
+2.0.0 (unreleased)
+==================
+
+- Fix compatibility with aiohttp>=0.21.0.
+- [routing] *Backwards-incompatible*: Renamed ``ResourceRouter.add_resource`` to ``ResourceRouter.add_resource_object``, to prevent clashing with aiohttp's URLDispatcher.
+
 1.0.0 (2015-10-27)
 ==================